backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
sentence-transformers
faiss-cpu
openai
python-dotenv
huggingface-hub
sentence-transformers